+++ DawnLib/src/API/TerminalCommands/DawnQueryCommandInfo.cs
@@ -5,7 +5,7 @@ public sealed class DawnQueryCommandInfo
 {
     public DawnTerminalCommandInfo ParentInfo { get; internal set; }
 
-    internal DawnQueryCommandInfo(Func<string> queryFunc, Func<string> cancelFunc, string continueKeyword, string cancelKeyword, FuncProvider<bool> continueProvider, DawnEvent? onQueryContinuedEvent)
+    internal DawnQueryCommandInfo(Func<string> queryFunc, Func<string> cancelFunc, string continueKeyword, string cancelKeyword, FuncProvider<bool> continueProvider, DawnEvent<bool>? onQueryContinuedEvent)
     {
         QueryFunc = queryFunc;
         CancelFunc = cancelFunc;
@@ -20,7 +20,7 @@ public sealed class DawnQueryCommandInfo
     internal string ContinueKeyword { get; }
     internal string CancelKeyword { get; }
     internal FuncProvider<bool> ContinueProvider { get; private set; }
-    internal DawnEvent? OnQueryContinuedEvent { get; private set; }
+    internal DawnEvent<bool>? OnQueryContinuedEvent { get; private set; }
 
     public TerminalNode? QueryNode { get; internal set; }
     public TerminalNode? CancelNode { get; internal set; }